requirements.txt
# requirements.txt

# Core
pandas
numpy
joblib
python-dotenv

# Modeling
scikit-learn
catboost

# Backend
fastapi
uvicorn

# Explanation
shap
requests
tqdm

# Frontend
streamlit
streamlit-shap

# Utils
loguru
typing-extensions